add sigil crafting and cleaned up creatables
