+++ src/lib/createables/barrows.ts
@@ -1,0 +1,186 @@
+import itemID from '../util/itemID';
+import { Bank } from '../types';
+
+interface Createable {
+	name: string;
+	outputItems: Bank;
+	inputItems: Bank;
+	smithingLevel?: number;
+	addOutputToCollectionLog?: boolean;
+	cantHaveItems: Bank;
+	firemakingLevel?: number;
+	craftingLevel?: number;
+	prayerLevel?: number;
+}
+const Barrows: Createable[] = [
+	{
+		name: "Verac's armour set",
+		inputItems: {
+			[itemID("Verac's helm")]: 1,
+			[itemID("Verac's brassard")]: 1,
+			[itemID("Verac's plateskirt")]: 1,
+			[itemID("Verac's flail")]: 1
+		},
+		outputItems: {
+			[itemID("Verac's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Veracs',
+		inputItems: {
+			[itemID("Verac's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Verac's helm")]: 1,
+			[itemID("Verac's brassard")]: 1,
+			[itemID("Verac's plateskirt")]: 1,
+			[itemID("Verac's flail")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Dharok's armour set",
+		inputItems: {
+			[itemID("Dharok's helm")]: 1,
+			[itemID("Dharok's platebody")]: 1,
+			[itemID("Dharok's platelegs")]: 1,
+			[itemID("Dharok's greataxe")]: 1
+		},
+		outputItems: {
+			[itemID("Dharok's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Dharoks',
+		inputItems: {
+			[itemID("Dharok's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Dharok's helm")]: 1,
+			[itemID("Dharok's platebody")]: 1,
+			[itemID("Dharok's platelegs")]: 1,
+			[itemID("Dharok's greataxe")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Guthan's armour set",
+		inputItems: {
+			[itemID("Guthan's helm")]: 1,
+			[itemID("Guthan's platebody")]: 1,
+			[itemID("Guthan's chainskirt")]: 1,
+			[itemID("Guthan's warspear")]: 1
+		},
+		outputItems: {
+			[itemID("Guthan's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Guthans',
+		inputItems: {
+			[itemID("Guthan's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Guthan's helm")]: 1,
+			[itemID("Guthan's platebody")]: 1,
+			[itemID("Guthan's chainskirt")]: 1,
+			[itemID("Guthan's warspear")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Ahrim's armour set",
+		inputItems: {
+			[itemID("Ahrim's hood")]: 1,
+			[itemID("Ahrim's robetop")]: 1,
+			[itemID("Ahrim's robeskirt")]: 1,
+			[itemID("Ahrim's staff")]: 1
+		},
+		outputItems: {
+			[itemID("Ahrim's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Ahrims',
+		inputItems: {
+			[itemID("Ahrim's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Ahrim's hood")]: 1,
+			[itemID("Ahrim's robetop")]: 1,
+			[itemID("Ahrim's robeskirt")]: 1,
+			[itemID("Ahrim's staff")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Torag's armour set",
+		inputItems: {
+			[itemID("Torag's helm")]: 1,
+			[itemID("Torag's platebody")]: 1,
+			[itemID("Torag's platelegs")]: 1,
+			[itemID("Torag's hammers")]: 1
+		},
+		outputItems: {
+			[itemID("Torag's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Torags',
+		inputItems: {
+			[itemID("Torag's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Torag's helm")]: 1,
+			[itemID("Torag's platebody")]: 1,
+			[itemID("Torag's platelegs")]: 1,
+			[itemID("Torag's hammers")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Karil's armour set",
+		inputItems: {
+			[itemID("Karil's coif")]: 1,
+			[itemID("Karil's leathertop")]: 1,
+			[itemID("Karil's leatherskirt")]: 1,
+			[itemID("Karil's crossbow")]: 1
+		},
+		outputItems: {
+			[itemID("Karil's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Karils',
+		inputItems: {
+			[itemID("Karil's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Karil's coif")]: 1,
+			[itemID("Karil's leathertop")]: 1,
+			[itemID("Karil's leatherskirt")]: 1,
+			[itemID("Karil's crossbow")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	}
+];
+
+export default Barrows;
